pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conclaw"
version = "0.1.0"
description = "CLI-based agent for automating Word and Excel file operations"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
conclaw = "conclaw.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/conclaw"]
